backend/requirements.txt
asgiref
Django
django-filter==25.1
djangorestframework==3.16.0
drf-yasg==1.21.10
Faker==37.1.0
inflection
Markdown==3.8
mysqlclient==2.2.7
packaging
PyMySQL==1.1.1
python-dotenv==1.1.0
pytz
PyYAML
setuptools==75.8.0
sqlparse
typing_extensions
tzdata==2025.2
uritemplate
wheel==0.45.1
